Refactor Details-view
