--- src/assets/css/styles.css
@@ -1,198 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;700&display=swap");
-* {
-  margin: 0px;
-  padding: 0px;
-  box-sizing: border-box;
-}
-
-body {
-  font-family: "Roboto", sans-serif;
-  position: relative;
-  background-color: #272727;
-}
-
-body::before {
-  content: "";
-  position: absolute;
-  margin: 0 auto;
-  width: 65vw;
-  min-height: 100%;
-  left: 0px;
-  right: 0px;
-  top: 15rem;
-  z-index: -100;
-  background: linear-gradient(90deg, rgba(255, 0, 245, 0.7) 0%, rgba(0, 128, 255, 0.918) 100%);
-  filter: blur(90px);
-}
-
-.container {
-  width: 95%;
-  margin: 0 auto;
-  max-width: 1400px;
-}
-
-.header {
-  background-color: #272727;
-}
-
-.container-header {
-  display: flex;
-  justify-content: space-between;
-  min-height: 4rem;
-  align-items: center;
-  color: white;
-}
-.container-header .logo {
-  font-size: 1.5rem;
-}
-
-.nav {
-  display: flex;
-  flex-direction: column;
-}
-.nav .fa-bars, .nav .fa-times {
-  height: 3rem;
-  width: 2rem;
-  align-self: flex-end;
-}
-
-.container-nav {
-  position: absolute;
-  top: 4rem;
-  right: 0px;
-  left: 0px;
-  width: 100vw;
-  padding: 2rem;
-  display: flex;
-  flex-direction: column;
-  background-color: #272727;
-}
-
-.filmdetails--main-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 5%;
-  margin-top: 20px;
-}
-.filmdetails--main-container .filmdetails--title {
-  color: white;
-  font-size: 50px;
-  margin-left: -10px;
-  padding: 10px 0px;
-  font-weight: bolder;
-}
-
-.filmdetails--stars .star, .filmdetails--stars .star-point {
-  cursor: pointer;
-  text-shadow: 0 0 31px inset grey;
-  background-color: transparent;
-  border: 1px grey;
-  outline: none;
-  color: white;
-  width: 20px;
-  height: 30px;
-}
-.filmdetails--stars .star:hover, .filmdetails--stars .star-point:hover {
-  color: yellow;
-}
-.filmdetails--stars .star-point {
-  color: yellow;
-}
-.filmdetails--stars .star-point:hover {
-  color: white;
-}
-
-.filmdetails--show-or-not {
-  width: 20%;
-}
-.filmdetails--show-or-not .filmdetails--check {
-  background-color: transparent;
-  border: 0px;
-  outline: none;
-  height: 30px;
-}
-
-.filmdetails--section {
-  display: flex;
-  flex-direction: row;
-  gap: 10px;
-  flex-wrap: wrap;
-  align-items: center;
-  justify-content: center;
-  text-align: center;
-  border-radius: 20px;
-  background-color: rgba(255, 255, 255, 0.541);
-  box-shadow: 0 0 5px grey, 0 0 4px inset white;
-  margin: 20px;
-  padding: 20px 10px;
-  width: 70%;
-}
-@media (max-width: 760px) {
-  .filmdetails--section {
-    width: 90%;
-  }
-}
-
-.filmdetails--valorations {
-  justify-content: space-around;
-}
-
-.filmdetails--frame-photo {
-  width: 500px;
-  height: 500px;
-  border-radius: 20px;
-  overflow: hidden;
-  margin-bottom: 20px;
-}
-@media (max-width: 580px) {
-  .filmdetails--frame-photo {
-    width: 250px;
-    height: 250px;
-  }
-}
-.filmdetails--frame-photo img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-
-.filmdetails--description, .filmdetails--cast {
-  margin: 20px;
-}
-
-.filmdetails--cast {
-  display: flex;
-  flex-direction: row;
-  gap: 20px;
-}
-.filmdetails--cast .castcomponent--container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 20px;
-  width: 200px;
-  height: 250px;
-  border-radius: 20px;
-  background-color: rgba(206, 199, 199, 0.603);
-  box-shadow: 0 0 4px inset white, 0 0 5px rgba(206, 198, 198, 0.445);
-}
-.filmdetails--cast .castcomponent--container .castcomponent--frame-photo {
-  border-radius: 50%;
-  overflow: hidden;
-  margin: 30px;
-  width: 100px;
-  height: 100px;
-}
-.filmdetails--cast .castcomponent--container .castcomponent--frame-photo img {
-  width: 100%;
-  height: 100%;
-  margin: auto;
-}
-
-.filmdetails--presentation {
-  display: flex;
-  flex-direction: column;
-}
-
-/*# sourceMappingURL=styles.css.map */
