+++ src/components/FilmDetails/_filmDetails.scss
@@ -16,15 +16,19 @@
 
 
 .filmdetails--stars{
+    display: flex;
+    flex-direction: row;
+    gap: 10px;
+    width: 300px;
 
     .star, .star-point{
         cursor: pointer;
-        text-shadow: 0 0 31px inset grey;
+        text-shadow: 0 0 31px inset rgb(143, 60, 60);
         background-color: transparent;
         border: 1px grey;
         outline: none;
         color: white;
-        width: 20px;
+        width: 50%;
         height: 30px;
 
         &:hover{
@@ -34,10 +38,6 @@
 
     .star-point{
         color: yellow;
-
-        &:hover{
-            color:white;
-        }
     }
 }
 
@@ -45,6 +45,7 @@
     width: 20%;
 
     .filmdetails--check{
+        cursor: pointer;
         background-color: transparent;
         border: 0px;
         outline: none;
