--- src/views/random
@@ -1,1 +1,0 @@
-ahhhhhhhhh
